fix: address review feedback (v2 test API, redundant conversions, user.home guard)
- guard null user.home when resolving the desktop theme file
- migrate :design Compose UI tests to the v2 runComposeUiTest API
- drop redundant .toDouble() in the contrast luminance helper
- use the value from shouldNotBeNull() instead of !! in the typography test

+++ design/src/commonTest/kotlin/com/yonatankarp/agentdesk/design/token/AgentDeskContrastTest.kt
@@ -42,7 +42,7 @@ private fun relativeLuminance(color: Color): Double {
     fun linearize(channel: Float): Double = if (channel <= 0.03928f) {
         channel / 12.92
     } else {
-        ((channel + 0.055) / 1.055).toDouble().pow(2.4)
+        ((channel + 0.055) / 1.055).pow(2.4)
     }
     return 0.2126 * linearize(color.red) + 0.7152 * linearize(color.green) + 0.0722 * linearize(color.blue)
 }
